Remove AsJSON middleware
